requirements.txt
google-generativeai>=0.7.2
python-dotenv>=1.0.1
typer>=0.12.5
flask>=3.0.0
